fix(filter): drop param fallback, reject via schema
- Remove parsePositiveIntParam/parseNonNegativeIntParam and MAX_* caps;
  new toNumberParam only defaults absent params, bad values pass through
- NaN/out-of-range/decimal params now fail schema validation and render
  the full-page error instead of being silently corrected
- Add .int() to page/minRatingCount schemas so decimals hit validation
  instead of throwing from Prisma skip
- Replace parser tests with toNumberParam and schema hard-reject cases

+++ validations/resource.ts
@@ -7,6 +7,6 @@ export const resourceSchema = z.object({
     z.literal('like')
   ]),
   sortOrder: z.union([z.literal('asc'), z.literal('desc')]),
-  page: z.coerce.number().min(1).max(9999999),
+  page: z.coerce.number().int().min(1).max(9999999),
   limit: z.coerce.number().min(1).max(50)
 })
